+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fixed handling of `in` environment markers with commas.
 - Fixed a `UnicodeDecodeError` when an error occurs in venv.
+- Fixed Python requirements not properly set when resolving dependencies.
 
 
 ## [0.10.1] - 2018-05-28